fix typo, kes++
Note that this also is the first commit on the host, which means a fresh
install of statocles. It seems to have bumped quite a bit of metadata
and a few page urls to include trailing slashes. If that ends up being a
problem then we know where to look.