add "Day N" to the 2018 entry titles
Originally I left them out because when looking at the site it just
seemed to add clutter (vs last year when I had them). That said when
linking inbound from sites like reddit or twitter (where titles are used
to build content and the advent nature isn't obvious) having the day
portion is informative to the reader just in the title.