add extensible/reusable controller posts
These two posts get in to detail about writing reusable controllers and
making those controllers inheritable (in proper OO style).

I think if I have multiple posts in a series, I'm going to do them
weekly, and try to one "series" (which may only be one post) monthly.
That said, I have another post in this series to make: Extending
controllers with Roles (which sounds like it'll be a lot of fun).